requirements.txt
openai>=1.55.3
python-dotenv==1.0.1
google-api-python-client==2.118.0
beautifulsoup4==4.12.3
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
crawl4ai>=0.5.0
requests==2.31.0
pytest==7.4.3
pyOpenSSL>=23.0.0